+++ tests/test_hyperparams_opt.py
@@ -6,7 +6,7 @@ import pytest
 
 
 def _assert_eq(left, right):
-    assert left == right, '{} != {}'.format(left, right)
+    assert left == right, f'{left} != {right}'
 
 
 N_STEPS = 100
@@ -21,7 +21,7 @@ experiments = {}
 
 for algo in ALGOS:
     for env_id in ENV_IDS:
-        experiments['{}-{}'.format(algo, env_id)] = (algo, env_id)
+        experiments[f'{algo}-{env_id}'] = (algo, env_id)
 
 # Test for SAC
 experiments['sac-Pendulum-v0'] = ('sac', 'Pendulum-v0')
@@ -38,22 +38,21 @@ if os.path.isdir(LOG_FOLDER):
 @pytest.mark.parametrize("experiment", experiments.keys())
 def test_optimize(sampler, pruner, experiment):
     algo, env_id = experiments[experiment]
-    custom_params = ['n_steps:10'] if algo == 'ppo' else []
-    args = [
-        '-n', str(N_STEPS),
-        '--algo', algo,
-        '--env', env_id,
-        '-params', 'policy_kwargs:"dict(net_arch=[32])"', 'n_envs:1']\
-        + custom_params +\
-        ['--log-folder', LOG_FOLDER,
-        '--n-trials', str(N_TRIALS),
-        '--n-jobs', str(N_JOBS),
-        '--sampler', sampler,
-        '--pruner', pruner,
-        '--n-evaluations', str(2),
-        '--n-startup-trials', str(1),
-        '-optimize'
-    ]
+    args = ['-n', str(N_STEPS),
+            '--algo', algo,
+            '--env', env_id,
+            '-params', 'policy_kwargs:"dict(net_arch=[32])"', 'n_envs:1'
+            ]
+    args += ['n_steps:10'] if algo == 'ppo' else []
+    args += ['--log-folder', LOG_FOLDER,
+             '--n-trials', str(N_TRIALS),
+             '--n-jobs', str(N_JOBS),
+             '--sampler', sampler,
+             '--pruner', pruner,
+             '--n-evaluations', str(2),
+             '--n-startup-trials', str(1),
+             '-optimize'
+             ]
 
     return_code = subprocess.call(['python', 'train.py'] + args)
     _assert_eq(return_code, 0)