+++ train.py
@@ -9,12 +9,11 @@ from pprint import pprint
 
 import yaml
 import gym
-from gym.wrappers import AtariPreprocessing
 import seaborn
 import numpy as np
 import torch as th
 # For custom activation fn
-import torch.nn as nn  # pylint: disable=unused-import
+import torch.nn as nn  # noqa: F401 pytype: disable=unused-import
 
 from stable_baselines3.common.utils import set_random_seed
 # from stable_baselines3.common.cmd_util import make_atari_env
@@ -25,7 +24,7 @@ from stable_baselines3.common.utils import constant_fn
 from stable_baselines3.common.callbacks import CheckpointCallback, EvalCallback
 
 # Register custom envs
-import utils.import_envs  # pytype: disable=import-error
+import utils.import_envs  # noqa: F401 pytype: disable=import-error
 from utils import make_env, ALGOS, linear_schedule, get_latest_run_id, get_wrapper_class
 from utils.hyperparams_opt import hyperparam_optimization
 from utils.callbacks import SaveVecNormalizeCallback
@@ -34,7 +33,7 @@ from utils.utils import StoreDict, get_callback_class
 
 seaborn.set()
 
-if __name__ == '__main__':
+if __name__ == '__main__':  # noqa: C901
     parser = argparse.ArgumentParser()
     parser.add_argument('--algo', help='RL Algorithm', default='ppo',
                         type=str, required=False, choices=list(ALGOS.keys()))
@@ -109,7 +108,7 @@ if __name__ == '__main__':
 
     set_random_seed(args.seed)
 
-    # Setting num threads to 1 by default, it makes everything run faster
+    # Setting num threads to 1 makes things run faster on cpu
     if args.num_threads > 0:
         if args.verbose > 1:
             print(f"Setting torch.num_threads to {args.num_threads}")
@@ -220,22 +219,25 @@ if __name__ == '__main__':
         del hyperparams['callback']
 
     if args.save_freq > 0:
+        # Account for the number of parallel environments
+        args.save_freq = max(args.save_freq // n_envs, 1)
         callbacks.append(CheckpointCallback(save_freq=args.save_freq,
                                             save_path=save_path, name_prefix='rl_model', verbose=1))
 
-
-    def create_env(n_envs, eval_env=False):
+    def create_env(n_envs, eval_env=False, no_log=False):
         """
         Create the environment and wrap it if necessary
         :param n_envs: (int)
         :param eval_env: (bool) Whether is it an environment used for evaluation or not
+        :param no_log: (bool) Do not log training when doing hyperparameter optim
+            (issue with writing the same file)
         :return: (Union[gym.Env, VecEnv])
         """
         global hyperparams
         global env_kwargs
 
         # Do not log eval env (issue with writing the same file)
-        log_dir = None if eval_env else save_path
+        log_dir = None if eval_env or no_log else save_path
 
         if n_envs == 1:
             env = DummyVecEnv([make_env(env_id, 0, args.seed,
@@ -247,12 +249,21 @@ if __name__ == '__main__':
             env = DummyVecEnv([make_env(env_id, i, args.seed, log_dir=log_dir, env_kwargs=env_kwargs,
                                         wrapper_class=env_wrapper) for i in range(n_envs)])
         if normalize:
+            # Copy to avoid changing default values by reference
+            local_normalize_kwargs = normalize_kwargs.copy()
+            # Do not normalize reward for env used for evaluation
+            if eval_env:
+                if len(local_normalize_kwargs) > 0:
+                    local_normalize_kwargs['norm_reward'] = False
+                else:
+                    local_normalize_kwargs = {'norm_reward': False}
+
             if args.verbose > 0:
-                if len(normalize_kwargs) > 0:
-                    print(f"Normalization activated: {normalize_kwargs}")
+                if len(local_normalize_kwargs) > 0:
+                    print(f"Normalization activated: {local_normalize_kwargs}")
                 else:
                     print("Normalizing input and reward")
-            env = VecNormalize(env, **normalize_kwargs)
+            env = VecNormalize(env, **local_normalize_kwargs)
 
         # Optional Frame-stacking
         if hyperparams.get('frame_stack', False):
@@ -266,12 +277,11 @@ if __name__ == '__main__':
             env = VecTransposeImage(env)
         return env
 
-
     env = create_env(n_envs)
 
     # Create test env if needed, do not normalize reward
     eval_env = None
-    if args.eval_freq > 0:
+    if args.eval_freq > 0 and not args.optimize_hyperparameters:
         # Account for the number of parallel environments
         args.eval_freq = max(args.eval_freq // n_envs, 1)
 
@@ -284,15 +294,6 @@ if __name__ == '__main__':
                                          log_path=save_path, eval_freq=args.eval_freq)
             callbacks.append(eval_callback)
         else:
-            # Do not normalize the rewards of the eval env
-            old_kwargs = None
-            if normalize:
-                if len(normalize_kwargs) > 0:
-                    old_kwargs = normalize_kwargs.copy()
-                    normalize_kwargs['norm_reward'] = False
-                else:
-                    normalize_kwargs = {'norm_reward': False}
-
             if args.verbose > 0:
                 print("Creating test environment")
 
@@ -302,10 +303,6 @@ if __name__ == '__main__':
                                          log_path=save_path, eval_freq=args.eval_freq,
                                          deterministic=not is_atari)
             callbacks.append(eval_callback)
-
-            # Restore original kwargs
-            if old_kwargs is not None:
-                normalize_kwargs = old_kwargs.copy()
 
     # TODO: check for hyperparameters optimization
     # TODO: check What happens with the eval env when using frame stack
@@ -374,14 +371,12 @@ if __name__ == '__main__':
         if args.verbose > 0:
             print("Optimizing hyperparameters")
 
-
         def create_model(*_args, **kwargs):
             """
             Helper to create a model with different hyperparameters
             """
-            return ALGOS[args.algo](env=create_env(n_envs, eval_env=True), tensorboard_log=tensorboard_log,
+            return ALGOS[args.algo](env=create_env(n_envs, no_log=True), tensorboard_log=tensorboard_log,
                                     verbose=0, **kwargs)
-
 
         data_frame = hyperparam_optimization(args.algo, create_model, create_env, n_trials=args.n_trials,
                                              n_timesteps=n_timesteps, hyperparams=hyperparams,