+++ utils/utils.py
@@ -7,13 +7,13 @@ from typing import Dict, Tuple
 
 import gym
 # For custom activation fn
-import torch.nn as nn  # pylint: disable=unused-import
+import torch.nn as nn  # noqa: F401 pylint: disable=unused-import
 
 from stable_baselines3.common.monitor import Monitor
 # from stable_baselines3.common import logger
 from stable_baselines3 import A2C, PPO, SAC, TD3
 from stable_baselines3.common.vec_env import (DummyVecEnv, VecNormalize,
-                                             VecFrameStack, SubprocVecEnv)
+                                              VecFrameStack, SubprocVecEnv)
 # from stable_baselines3.common.cmd_util import make_atari_env
 from stable_baselines3.common.utils import set_random_seed
 
@@ -140,8 +140,8 @@ def get_callback_class(hyperparams):
             # Handle keyword arguments
             if isinstance(callback_name, dict):
                 assert len(callback_name) == 1, ("You have an error in the formatting "
-                                                f"of your YAML file near {callback_name}. "
-                                                "You should check the indentation.")
+                                                 f"of your YAML file near {callback_name}. "
+                                                 "You should check the indentation.")
                 callback_dict = callback_name
                 callback_name = list(callback_dict.keys())[0]
                 kwargs = callback_dict[callback_name]