+++ CHANGELOG.md
@@ -4,14 +4,21 @@
 
 ### New Features
 - Distributed optimization (@SammyRamone)
+- Added ``--load-checkpoints`` to load particular checkpoints
+- Added ``--num-threads`` to enjoy script
 
 ### Bug fixes
-- Fix optuna warning (@SammyRamone)
+- Fixed optuna warning (@SammyRamone)
+- Fixed `--save-freq` which was not taking parallel env into account
+- Use 'maximize' for hyperparameter optimization (@SammyRamone)
+- Fixed a bug where reward where not normalized when doing hyperparameter optimization (@caburu)
 
 ### Documentation
 
 ### Other
+- Reformat `enjoy.py`, `test_enjoy.py`, `test_hyperparams_opt.py`, `test_train.py`, `train.py`, `callbacks.py`, `hyperparams_opt.py`, `utils.py`, `wrappers.py` (@salmannotkhan)
 - Reformat `record_video.py` (@salmannotkhan)
 - Added codestyle check `make lint` using flake8
 - Reformat `benchmark.py` (@salmannotkhan)
 - Added github ci
+- Fixes most linter warnings